Added elo range chart

+++ db/migrate/20170414234601_add_current_elo_to_game_players.rb
@@ -1,5 +1,5 @@
 class AddCurrentEloToGamePlayers < ActiveRecord::Migration[5.0]
   def change
-    add_column :game_players, :current_elo, :float
+    add_column :game_players, :current_elo, :integer
   end
 end
